Write unit tests for the proxy package
OnError should be different for UIRequest and for GitHub webhook
request
Move test to validate session to ui request tests as that's the only
place where it can occur
